backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.13.1
stripe==8.0.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
email-validator==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
requests==2.31.0 